requirements.txt
streamlit==1.32.2
python-dotenv==1.0.1
torch==2.2.2

# PDF Processing
pymupdf==1.23.26
pdfplumber==0.10.3

# LangChain Modern Split
langchain==0.1.17
langchain-core==0.1.52
langchain-community==0.0.37
langchain-text-splitters==0.0.1

# Embeddings & Vector DB
sentence-transformers==2.2.2
faiss-cpu==1.7.4
rank-bm25==0.2.2

# Ollama + HTTP
ollama==0.1.8
requests==2.31.0
